+++ packages/core/src/pubsub/incoming-channel.ts
@@ -67,6 +67,7 @@ export class IncomingChannel extends Observable<IPFSPubsubMessage> {
     readonly ipfs: IpfsApi,
     readonly topic: string,
     readonly resubscribeEvery: number,
+    readonly lateMessageAfter: number,
     readonly pubsubLogger: ServiceLogger,
     readonly logger: DiagnosticsLogger,
     readonly tasks: TaskQueue = new TaskQueue()
@@ -74,6 +75,7 @@ export class IncomingChannel extends Observable<IPFSPubsubMessage> {
     super((subscriber) => {
       new PubsubIncoming(ipfs, topic, pubsubLogger, logger, this.tasks)
         .pipe(
+          checkSlowObservable(lateMessageAfter, logger, 'IPFS did not provide any messages, please check your IPFS configuration.'),
           retryWhen((errors) =>
             errors.pipe(
               tap((e) => logger.err(e)),
@@ -101,5 +103,30 @@ export function filterExternal(
     filter((data) => data.isOuter),
     // remove the data container object from the observable chain
     map((data) => data.entry)
+  )
+}
+
+export function checkSlowObservable(
+  delay: number,
+  logger: DiagnosticsLogger,
+  description: string,
+): MonoTypeOperatorFunction<IPFSPubsubMessage> {
+  const createTimeout = () => {
+    return setTimeout(
+      () => {
+        logger.warn(`Message was not timely. ${description}`)
+      },
+      delay
+    )
+  }
+  let outstanding = createTimeout()
+  return pipe(
+    tap({
+      next: () => {
+        clearTimeout(outstanding)
+        outstanding = createTimeout()
+      },
+      complete: () => clearTimeout(outstanding)
+    })
   )
 }