+++ packages/anchor-listener/CHANGELOG.md
@@ -1,4 +1,44 @@
 # Change Log
 
+All notable changes to this project will be documented in this file.
+See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
+
+# [1.2.0](/compare/@ceramicnetwork/anchor-listener@1.2.0-rc.1...@ceramicnetwork/anchor-listener@1.2.0) (2023-01-23)
+
+**Note:** Version bump only for package @ceramicnetwork/anchor-listener
+
+
+
+
+
+# [1.2.0-rc.1](/compare/@ceramicnetwork/anchor-listener@1.2.0-rc.0...@ceramicnetwork/anchor-listener@1.2.0-rc.1) (2023-01-20)
+
+**Note:** Version bump only for package @ceramicnetwork/anchor-listener
+
+
+
+
+
+# [1.2.0-rc.0](https://github.com/ceramicnetwork/js-ceramic/compare/@ceramicnetwork/anchor-listener@1.1.0-rc.0...@ceramicnetwork/anchor-listener@1.2.0-rc.0) (2023-01-18)
+
+**Note:** Version bump only for package @ceramicnetwork/anchor-listener
+
+
+
+
+
+# 1.1.0-rc.0 (2023-01-11)
+
+
+### Features
+
+* **anchor-listener,anchor-utils:** Implement anchors listener and utils packages ([d9f4390](https://github.com/ceramicnetwork/js-ceramic/commit/d9f4390ef02ee44a31b399b94a805e64883932d5))
+
+
+
+
+
+# Change Log
+
 All notable changes to this project will be documented in this file. See
 [Conventional Commits](https://conventionalcommits.org) for commit guidelines.