+++ packages/anchor-utils/CHANGELOG.md
@@ -1,4 +1,28 @@
 # Change Log
 
+All notable changes to this project will be documented in this file.
+See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
+
+# [1.1.0](/compare/@ceramicnetwork/anchor-utils@1.1.0-rc.0...@ceramicnetwork/anchor-utils@1.1.0) (2023-01-23)
+
+**Note:** Version bump only for package @ceramicnetwork/anchor-utils
+
+
+
+
+
+# 1.1.0-rc.0 (2023-01-11)
+
+
+### Features
+
+* **anchor-listener,anchor-utils:** Implement anchors listener and utils packages ([d9f4390](https://github.com/ceramicnetwork/js-ceramic/commit/d9f4390ef02ee44a31b399b94a805e64883932d5))
+
+
+
+
+
+# Change Log
+
 All notable changes to this project will be documented in this file. See
 [Conventional Commits](https://conventionalcommits.org) for commit guidelines.