+++ packages/pinning-aggregation/CHANGELOG.md
@@ -3,6 +3,38 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [2.10.0](/compare/@ceramicnetwork/pinning-aggregation@2.10.0-rc.1...@ceramicnetwork/pinning-aggregation@2.10.0) (2023-01-23)
+
+**Note:** Version bump only for package @ceramicnetwork/pinning-aggregation
+
+
+
+
+
+# [2.10.0-rc.1](/compare/@ceramicnetwork/pinning-aggregation@2.10.0-rc.0...@ceramicnetwork/pinning-aggregation@2.10.0-rc.1) (2023-01-20)
+
+**Note:** Version bump only for package @ceramicnetwork/pinning-aggregation
+
+
+
+
+
+# [2.10.0-rc.0](https://github.com/ceramicnetwork/js-ceramic/compare/@ceramicnetwork/pinning-aggregation@2.9.0-rc.0...@ceramicnetwork/pinning-aggregation@2.10.0-rc.0) (2023-01-18)
+
+**Note:** Version bump only for package @ceramicnetwork/pinning-aggregation
+
+
+
+
+
+# [2.9.0-rc.0](https://github.com/ceramicnetwork/js-ceramic/compare/@ceramicnetwork/pinning-aggregation@2.8.0...@ceramicnetwork/pinning-aggregation@2.9.0-rc.0) (2023-01-11)
+
+**Note:** Version bump only for package @ceramicnetwork/pinning-aggregation
+
+
+
+
+
 # [2.8.0](https://github.com/ceramicnetwork/js-ceramic/compare/@ceramicnetwork/pinning-aggregation@2.8.0-rc.0...@ceramicnetwork/pinning-aggregation@2.8.0) (2023-01-05)
 
 **Note:** Version bump only for package @ceramicnetwork/pinning-aggregation